商品削除機能２

+++ app/controllers/items_controller.rb
@@ -30,12 +30,14 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
-   else
-    render :show
-   end
-
+    if user_signed_in? && @item.user_id == current_user.id 
+    
+      if @item.destroy
+        redirect_to root_path
+      else
+        render :show
+      end
+    end
   end
 
   def update
